Added validation and only deleting asset if exists(scenario-working on two tabs at same time).

+++ src/main/java/com/project/service/AssetService.java
@@ -32,6 +32,9 @@ public class AssetService {
 
     // Method to delete the asset by their id
     public void deleteAsset(int id) {
+        if(!assetRepository.findById(id).isPresent()){
+            throw new NotFoundException("Asset not found");
+        }
         assetRepository.deleteById(id);
     }
